double precision of nbody is added

+++ trunk/examples/oclNbody/inc/oclBodySystemCpu_dp.h
@@ -12,7 +12,7 @@
 #ifndef __CL_BODYSYSTEMCPU_H__
     #define __CL_BODYSYSTEMCPU_H__
 
-    #include "oclBodySystem.h"
+    #include "oclBodySystem_dp.h"
 
     // CPU Body System
     class BodySystemCPU : public BodySystem
@@ -21,13 +21,13 @@
             BodySystemCPU(int numBodies);
             virtual ~BodySystemCPU();
 
-            virtual void update(float deltaTime);
+            virtual void update(double deltaTime);
 
-            virtual void setSoftening(float softening) { m_softeningSquared = softening * softening; }
-            virtual void setDamping(float damping)     { m_damping = damping; }
+            virtual void setSoftening(double softening) { m_softeningSquared = softening * softening; }
+            virtual void setDamping(double damping)     { m_damping = damping; }
 
-            virtual float* getArray(BodyArray array);
-            virtual void   setArray(BodyArray array, const float* data);
+            virtual double* getArray(BodyArray array);
+            virtual void   setArray(BodyArray array, const double* data);
 
             virtual size_t getCurrentReadBuffer() const { return m_currentRead; }
 
@@ -38,15 +38,15 @@
             virtual void _finalize();
 
             void _computeNBodyGravitation();
-            void _integrateNBodySystem(float deltaTime);
+            void _integrateNBodySystem(double deltaTime);
             
         protected: // data
-            float* m_pos[2];
-            float* m_vel[2];
-            float* m_force;
+            double* m_pos[2];
+            double* m_vel[2];
+            double* m_force;
 
-            float m_softeningSquared;
-            float m_damping;
+            double m_softeningSquared;
+            double m_damping;
 
             unsigned int m_currentRead;
             unsigned int m_currentWrite;
